add navigation + home page
