+++ package.json
@@ -9,6 +9,8 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.1",
+    "@react-navigation/native-stack": "^6.9.7",
     "expo": "^47.0.10",
     "expo-gl": "^12.0.1",
     "expo-status-bar": "~1.4.2",
@@ -16,6 +18,8 @@
     "react": "18.1.0",
     "react-dom": "18.1.0",
     "react-native": "0.70.5",
+    "react-native-safe-area-context": "4.4.1",
+    "react-native-screens": "~3.18.0",
     "react-native-web": "~0.18.9",
     "standard": "^17.0.0",
     "three": "^0.148.0"
