+++ pages/Home.tsx
@@ -1,0 +1,20 @@
+import { View, Text, StyleSheet } from 'react-native';
+
+const Home = ({ navigation }) => {
+  return (
+    <View style={styles.container}>
+      <Text>Home page</Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center'
+  }
+});
+
+export default Home;
